+++ sciolyid/web/functions/images.py
@@ -1,49 +1,19 @@
 import csv
 import imghdr
 import os
-import shutil
 from typing import Dict, Optional, Set, Union
 
 import imagehash
 import requests
+from flask import abort
 from PIL import Image
 
 import sciolyid.config as config
-import sciolyid.web.tasks.git_tasks as git_tasks
 from sciolyid.web.config import logger
-from sciolyid.web.git import verify_repo
 
 VALID_MIMETYPES = ("image/jpeg", "image/png")
 VALID_IMG_TYPES = ("jpeg", "png")
 MAX_FILESIZE = 4000000  # 4 mb
-
-
-def add_images(
-    sources: list,
-    destinations: Union[str, list],
-    user_id: int,
-    username: str,
-    use_filenames: bool = True,
-):
-    different_dests = False
-    if isinstance(destinations, list):
-        if len(destinations) != len(sources):
-            logger.info("source/dest invalid")
-            raise IndexError("sources and destinations are not the same length")
-        different_dests = True
-
-    verify_repo.remote("origin").pull()
-    for i, item in enumerate(sources):
-        filename = item.split("/")[-1] if use_filenames else ""
-        destination_path = (
-            config.options["validation_local_dir"] + destinations[i]
-            if different_dests
-            else destinations
-        )
-        os.makedirs(destination_path, exist_ok=True)
-        shutil.copyfile(item, destination_path + filename)
-
-    git_tasks.push.delay(f"add images: id-{user_id}\n\nUsername: {username}", user_id)
 
 
 def find_duplicates(image, distance: int = 5, ignore_verify=False) -> list:
@@ -96,6 +66,26 @@ def generate_id_lookup() -> Optional[Dict[str, str]]:
     for filename, image_id in r:
         lookup[filename] = image_id
     return lookup
+
+
+def filename_lookup(start_path: str) -> dict:
+    id_lookup = generate_id_lookup()
+    if not id_lookup:
+        abort(500, "filename lookup failed!")
+    result = {}
+    stack = []
+    stack.append(start_path)
+    while stack:
+        current = stack.pop()
+        for child_filename in os.listdir(current):
+            child_path = current + "/" + child_filename
+            if os.path.isdir(child_path):
+                stack.append(child_path)
+                continue
+            if imghdr.what(child_path) in VALID_IMG_TYPES:
+                image_id = id_lookup["./" + os.path.relpath(child_path, start_path)]
+                result[image_id] = child_path
+    return result
 
 
 def verify_image(f, mimetype) -> Union[bool, str]:
